requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.10.4
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
numpy==2.2.3
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
yarl==1.18.3
SQLAlchemy==2.0.23
alembic==1.13.1
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3
aiosqlite==0.20.0
loguru==0.7.2
